cleaned code

+++ src/header_specific/inter_size.c
@@ -4,7 +4,7 @@ int inter_size(char* str, size_t len)
 {
     int size = 0, index = 0;
 
-    while (index < len)
+    while (index < (int)len)
     {
         size += str[index];
         index += 1;
